Add test for fullgraph compile
Graph breaks in model code are detrimental to performance when
models are compiled. Add a test for fullgraph compile of models.
For models where fullgraph capture works (fno, songunet), this test will
ensure that model code updates dont break it. For the other models
(healpix, graphcast) this test shows graphbreaks so they can be fixed.

Signed-off-by: Ramu Ramamurthy <ramu.ramamurthy@gmail.com>